Move OrdersVM sync action test to SyncActionUseCaseTests
